feat: Hide social reactions
When this feature is enabled, users will still be able to
like, share, or comment on a post. But they will not be able to
see how many likes (or other emojis) a post has, and in some
cases the number of shares a post has.

For the time being, users will be able to see the number of
comments a post has because hiding it will also hide the ability
to read comments.